+++ src/heuristic_functions.c
@@ -1,14 +1,14 @@
 #include "heuristic_functions.h"
 
-int dr_heuristic(cube_t* cube, ptable_data_t* p_data){
+size_t dr_heuristic(cube_t* cube, ptable_data_t* p_data){
     uint64_t      p1      = p_data->cube_to_index_func(cube, UD);
-    int pval_UD = (int)p_data->read_value_ptable_func(p1, p_data->ptable);
+    size_t pval_UD = p_data->read_value_ptable_func(p1, p_data->ptable);
     uint64_t      p2      = p_data->cube_to_index_func(cube, LR);
-    int  pval_LR = (int)p_data->read_value_ptable_func(p2, p_data->ptable);
+    size_t  pval_LR = p_data->read_value_ptable_func(p2, p_data->ptable);
     uint64_t      p3      = p_data->cube_to_index_func(cube, FB);
-    int  pval_FB = (int)p_data->read_value_ptable_func(p3, p_data->ptable);
+    size_t  pval_FB = p_data->read_value_ptable_func(p3, p_data->ptable);
     
-    int min = pval_UD;
+    size_t min = pval_UD;
     if (pval_FB < min){
         min = pval_FB;
     }
@@ -20,15 +20,15 @@ int dr_heuristic(cube_t* cube, ptable_data_t* p_data){
 }
 
 
-int htr_heuristic(cube_t* cube, ptable_data_t* p_data){
+size_t htr_heuristic(cube_t* cube, ptable_data_t* p_data){
     uint64_t      p1      = p_data->cube_to_index_func(cube, UD);
-    int pval_UD = (int)p_data->read_value_ptable_func(p1, p_data->ptable);
+    size_t pval_UD = p_data->read_value_ptable_func(p1, p_data->ptable);
     uint64_t      p2      = p_data->cube_to_index_func(cube, LR);
-    int  pval_LR = (int)p_data->read_value_ptable_func(p2, p_data->ptable);
+    size_t  pval_LR = p_data->read_value_ptable_func(p2, p_data->ptable);
     uint64_t      p3      = p_data->cube_to_index_func(cube, FB);
-    int  pval_FB = (int)p_data->read_value_ptable_func(p3, p_data->ptable);
+    size_t  pval_FB = p_data->read_value_ptable_func(p3, p_data->ptable);
     
-    int max = pval_UD;
+    size_t max = pval_UD;
     if (pval_FB > max){
         max = pval_FB;
     }
